Remove p6spy and datasource-proxy Spring Boot starters.
