--- sentry-spring-boot-starter-p6spy/api/sentry-spring-boot-starter-datasource-proxy.api
@@ -1,5 +1,0 @@
-public final class io/sentry/spring/boot/BuildConfig {
-	public static final field SENTRY_SPRING_BOOT_SDK_NAME Ljava/lang/String;
-	public static final field VERSION_NAME Ljava/lang/String;
-}
-
