+++ sentry-spring-boot-starter/api/sentry-spring-boot-starter.api
@@ -40,6 +40,10 @@ public class io/sentry/spring/boot/SentryProperties$Logging {
 	public fun setMinimumEventLevel (Lorg/slf4j/event/Level;)V
 }
 
+public final class io/sentry/spring/boot/SentryTracingCondition : org/springframework/boot/autoconfigure/condition/AnyNestedCondition {
+	public fun <init> ()V
+}
+
 public class io/sentry/spring/boot/datasource/dsproxy/SentryDsProxyAutoConfiguration {
 	public fun <init> ()V
 	public fun sentryQueryExecutionListener (Lio/sentry/IHub;)Lio/sentry/dsproxy/SentryQueryExecutionListener;
